Restructure the component folder structure
