+++ src/components/questions/Question.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import QuestionDetails from './QuestionDetails'
 import QuestionTags from './QuestionTags'
 
-import withCSS from '../../../lib/withCSS'
+import withCSS from '../../lib/withCSS'
 
 const Question = ({ head, id, lastUsed, tags, title, type, version }) =>
   (<div className="container">
@@ -69,4 +69,4 @@ Question.defaultProps = {
   version: 1,
 }
 
-export default withCSS(Question, ['segment'])
+export default Question
